refactor(cli): lism コマンド化と lism-ui 旧 bin の透過リダイレクト
- src/index.ts を `lism` CLI エントリに変更し、`lism ui {init,add,list}` のサブコマンド構造へ
- commands/{add,init,list}.ts を commands/ui/ 配下に移動、index.ts で ui コマンド構築
- createProgram.ts を分離し、index / index-legacy 両方から使う構造に
- bin/lism.mjs を新設（lism エントリ）、bin/cli.mjs は index-legacy.js を呼ぶ透過リダイレクトに
- index-legacy.ts は同一プロセス内で argv に 'ui' を挿入し、stderr に1行だけ rename 案内
