+++ packages/lism-cli/bin/cli.mjs
@@ -1,2 +1,2 @@
 #!/usr/bin/env node
-import '../dist/index.js';
+import '../dist/index-legacy.js';
