+++ packages/lism-cli/bin/lism.mjs
@@ -1,0 +1,2 @@
+#!/usr/bin/env node
+import '../dist/index.js';
