+++ packages/lism-cli/package.json
@@ -4,6 +4,7 @@
   "description": "CLI for adding Lism UI components to your project",
   "type": "module",
   "bin": {
+    "lism": "./bin/lism.mjs",
     "lism-ui": "./bin/cli.mjs"
   },
   "files": [
