+++ packages/lism-cli/tsup.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from 'tsup';
 
 export default defineConfig({
-  entry: ['src/index.ts'],
+  entry: ['src/index.ts', 'src/index-legacy.ts'],
   format: ['esm'],
   dts: false,
   clean: true,
